pyproject.toml
[project]
name = "gemini-mcp-server"
version = "0.1.0"
description = "MCP server providing Gemini AI web search and model calling capabilities"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "mcp>=1.6.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.0.0",
]
[project.scripts]
gemini-server = "gemini_mcp_server.__main__:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemini_mcp_server"]
force-include = { "gemini_mcp_server/tools.json" = "gemini_mcp_server/tools.json" }

[tool.hatch.build.targets.sdist]
include = [
  "gemini_mcp_server/**",
  "README.md",
  "LICENSE",
]
